Only stop workflow when npm audit fails with high or critical (#3)
Current moderate vulnerabilities arise from dependencies of tfjs-vis which is only used to display the model summary, not exposing the vulnerability to user. We could allow for a more granular control and only let tfjs-vis to have moderate but for the user case that is a bit of overkill. For the time being I will just allow moderate vulnerabilities on npm audit.